requirements.txt
--index-url https://upm.asx.com.au/repository/pypi-all/simple
--trusted-host upm.asx.com.au
#---------------
#paramiko==2.8.0
#JayDeBeApi==1.2.3
#-------------------
google-cloud-storage==2.1.0
#---------------
#google-cloud-logging==3.0.0
#google-cloud-secret-manager==2.8.0
#---------------
protobuf==3.20.1
apache-beam[gcp] # 2.46.0
#apache-beam[gcp]==2.34.0
pysmb
